git-svn-id: http://svn.seqan.de/seqan/trunk@1433 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/find_motif/find_motif_pmsp.h
@@ -455,7 +455,7 @@ pmsp(TStrings & result,
 					                     begin(dataset[seq_nr]), 
 										 end(dataset[seq_nr]), l, d, is_exact) )
 				{
-					int lower_limit = ceil(t*(model_type.threshold))-1;
+					int lower_limit = (int) ceil(t*(model_type.threshold))-1;
 					bool isMotif = true;
 					unsigned int number = 0;
 					for(typename Position< String< String<TPos2> > >::Type i=0; i<length(L); ++i)
